tf_serve/requirements.txt
tensorflow-serving-api
pillow
scipy
wget
numpy